+++ CHANGELOG.md
@@ -10,9 +10,11 @@ and this project adheres to
 
 ### Added
 
+- Support for optimizing source diagram using `--rewrite`
 - Skip only diagram wrapping with `--no-diagram-wrap`
 - Breaking of long elements over multiple lines in optional items `[]`
-- Plain text will now also be optimized when reasonable:
+- Plain text will now also be optimized when reasonable: Text will not be
+  optimized when using the `--no-optimizations` flag.
 
   ```
   [ [ a ] ] => [ a ]